Refactoring cauzzi_2014
